scripts/requirements.txt
--index-url https://pypi.org/simple

pillow==11.1.0